Перераспределение по файлам и придание адекватных имён
